hfp hf connection establishment

+++ src/hfp_ag.h
@@ -54,7 +54,7 @@ extern "C" {
 #endif
 
 void hfp_ag_create_service(uint8_t * service, int rfcomm_channel_nr, const char * name, uint8_t ability_to_reject_call, uint16_t supported_features);
-void hfp_ag_init(uint16_t rfcomm_channel_nr, uint16_t supported_features, uint8_t * codecs, int num_codecs);
+void hfp_ag_init(uint16_t rfcomm_channel_nr, uint32_t supported_features, uint8_t * codecs, int num_codecs);
 void hfp_ag_connect(bd_addr_t bd_addr);
 void hfp_ag_disconnect(bd_addr_t bd_addr);
